feat: Fetch data from react-router-dom function in Menu component for use in Order component
This commit introduces functionality to fetch data retrieved from the react-router-dom function in the Menu component and utilize it in the Order component. The useFetcher hook is employed for this purpose. The fetched menu data is then passed as a prop, containing ingredients for each pizza in the customer's cart, to be displayed in the OrderItem component. By incorporating this feature, the application enhances the user experience by providing comprehensive information about the ordered items.