feat: Allow editing of existing cabins in createCabinForm
This commit enhances the functionality of the createCabinForm component to support both creating a new cabin and editing an existing cabin. The component now accepts a prop called 'cabinToEdit', which contains the details of the cabin to be edited. We create a variable called 'editValues' that consists of all the content of the 'cabinToEdit' prop except the 'id'. Additionally, a variable called 'isEditSession' is introduced to indicate whether the createCabinForm component is being used to create a new cabin or to edit an existing one. We utilize the 'defaultValues' option of the useForm hook provided by react-hook-form, setting it to 'editValues' only if 'isEditSession' is true.