feat: Implement useMutation hook for editing cabins
This commit adds another useMutation hook to enable editing of existing cabins. The new createEditFunction is used along with the useMutation hook to edit a cabin. In this hook, the mutate function is set to editCabin. Previously, in the code for creating a new cabin, the mutate function was set to createCabin function. Both createCabin and editCabin functions are then used in the onSubmit handle function, which is responsible for submitting the necessary information to create or edit the cabin.